Added main python and config file
Running readMoisture.py can currently read analog input from the moisture sensor, compare it to the min/max values of the sensor, and report that data to a MySQL Server along with the DateTime it was recorded. The config file is to give users a single area to easily change variables that will be different from user to user.